Disable chrono default features. Fix RUSTSEC-2020-0071. (#16)
This should remove the complaint about RUSTSEC-2020-0071 being caused by the axum_session library. It may still appear due to the usage of default features in surrealdb lib, if surrealdb feature is enabled in this lib. I've also suggested they do this.

I ran `cargo fmt` and it added a newline to this as well.